final doc update before archiving repo
I don't see a need to maintain this extension any longer, chasing API
changes for what looks to be a minimally-modified copy of
Special:AncientPages and could now be reimplemented with hooks.

In its current state it doesn't work with MW 1.36+ due to $wgContLang
going away.  Updating that to MediaWikiServices::getInstance()->getContentLanguage()
makes it appear to work in 1.39, but no promises.